add ability to mutate pods based on annotation
This change adds ability to mark pod for mutation with
`cmk.intel.com/mutate` annotation.
This is useful for pods/conatiners which are not requesting
ERs (exclusive cores resources), but still want to use CMK
for running their processes on shared/infra cores.
Annotation values that will cause mutation include:
`"true"`, `"True"` and `"1"`.

Signed-off-by: Przemyslaw Lal <przemyslawx.lal@intel.com>