fix git remote HTTPS URL handling in make_release script
This change fixes an issue with the make_release script.
makre_release.py assumed that git remote is set to URL in SSH format.
This caused an issue when the repo was cloned using HTTPS URL.
This commit makes the relase script work for both HTTPS and SSH remote URLs.

Signed-off-by: Przemyslaw Lal <przemyslawx.lal@intel.com>