don't exit on exceptions during uninstall
This fixes error when not all resources are removed when running `cmk uninstall` command and trying to remove object that has been already deleted.
Instead, a warning message is displayed when the resource already doesn't exist and the uninstall script continues execution, trying to remove remaining resources in the best effort mode.